requirements.txt
requests==2.32.5
singer-python==6.3.0
backoff==2.2.1
